Power of number

+++ Recursion/Factorial/Factorial.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 
-int factorial(int n) {
+long int factorial(int n) {
     if (n == 0) {
         return 1;
     }
