+++ Recursion/Factorial/factorial.java
@@ -2,7 +2,7 @@ import java.util.Scanner;
 
 class Factorial {
 
-    static int factorial(int n) {
+    static long factorial(int n) {
         if (n == 0) {
             return 1;
         }
